RAG Documents Q&A/requirements.txt
langchain
langchain_core
langchain_community
langchain_groq
streamlit
python-dotenv
faiss-cpu
pypdf
langchain-ollama